backend/pyproject.toml
[project]
name = "okada-hackathon-rag-voice"
version = "0.1.0"
description = "A multimodal RAG voice agent for the Okada Hackathon."
authors = [{ name = "Tanmay", email = "tanmaysk22@gmail.com" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "pydantic>=2.7.1",
    "langchain-core>=0.2.0",
    "langgraph>=0.0.60",
    "redis>=5.0.4",
    "langchain-community>=0.0.38",
    "pypdf>=4.2.0",
    "docx2txt>=0.8",
    "unstructured[docx,pdf]>=0.12.6",
    "langchain-text-splitters>=0.2.0",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.2.0",
    "langchain-google-genai>=0.0.18",
    "tabulate>=0.9.0",
    "langchain-tavily>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.2.1",
    "google-cloud-speech>=2.25.0",
    "google-cloud-texttospeech>=2.17.0",
    "google-genai>=0.5.0",
    "tavily-python>=0.3.0",
    "websockets>=12.0",
    "langchain_experimental>=0.0.18",
    "tabulate",
    "llama-index>=0.10.40",
    "pandas>=2.2.2",
    "llama-index-llms-langchain",
    "llama-index-experimental",
    "llama-index-core"
]

[tool.uv]
# This space is reserved for any uv-specific configurations you might need later.